+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- [Common] Filter `Camera` alerts by road corridor: suppress alerts when user is on a parallel
+  road (cross-track distance > 40 m from camera axis)
 - [iOS] Bump min supported iOS 16
 - [iOS] Migrate from Cocoapods to SPM
 - [iOS] Update Google Maps to 10.13.0